lwm2m:  new test_for_change_SocketAddress add time for update (Redis)
